generate items moved to a new file
