+++ react-arcade/src/components/Game-Item.jsx
@@ -1,0 +1,10 @@
+const Item = ({ id, level }) => {
+    const className = `item level${level}`;
+    return (
+      <div className={className}>
+        {id}
+      </div>
+    );
+  };
+
+  export default Item
